Lab7_Python_LowCode/requirements.txt
pandas
python-dotenv
datasets
Pillow
pydantic
cohere